Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,10 +2,10 @@
 
 int main(void)
 {
-    float r;
+    int r;
     scanf("%f", r);
 
     float area = 3.14*(r**2);
     printf("Area: %f", area);
-    
+
 }
